Region lat/lng databinding via string
Currently, if you try and backspace in a lat/lng field on the region picker, the UI behaves oddly. Whatever the value of the field, it converts to a double and then the field is updated to this value. This means that attempts to backspace the decimal point fail (a double representation is always "x.y").

Rather than databinding to the double directly, this should go via a string method that does the double validation. This allows the user to momentarily input something that isn't a valid double without forcing an update to the UI. What's still to remain is to provide feedback to the user by setting field.error() if the number isn't a valid entry (valid and in range). Values are currently silently fixed or ignored if they're outside the range.

Also fixes part of issue #740 and allows negative numbers.