Fix bug with Geocoder resolve being called from wrong thread
This ended up being a bigger re-work of how the reverse geocoder is
called and how it works. There was a problem with the reverse geocoder
being passed a reference to a "target" entity that needed to be updated
with the result. The problem is that this is sometimes a View, and you
can't update a view from anything other than the UI thread. This fix
tries to solve this by using LiveData more appropriately, so that the
view doesn't need to be explicitly updated, but can react to data
changes.