Create the mapFragment explicitly in MapActivity onResume
It turns out that `FragmentManager.executePendingTransactions` calls
`Fragment.performCreateView`, which on OSM map then comes around and
calls `refreshMarkers`. This last method assumes that we have a handle
to the `MapFragment` being created, which we don't because we're still
executing pending transactions and haven't had a chance to grab it yet.
Then we get an uninitializedpropertyaccessexception.

So, we create the MapFragment explicitly from the factory, and pass that
through to the transaction.