Contacts Testing re-factor
One of the problems of the espresso contact tests was dealing with the
race between the test environment starting, configuring the MQTT
connection and then starting the test flows, versus the device just
pinging its current location to the broker (which creates a contact).

To get around this, we can now reset the ContactsRepo on every test with
a CLEAR_ALL intent (along with some idlingpolicy sync). Hopefully gives
better repeatable tests.